refactor: GameResult 객체를 컨트롤러로 옮겨 출력을 담당
