+++ src/test/java/blackjack/domain/ResultTest.java
@@ -1,9 +1,9 @@
 package blackjack.domain;
 
+import blackjack.controller.GameResultController;
+import java.util.Collections;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.Collections;
 
 public class ResultTest {
     @Test
@@ -11,7 +11,7 @@ public class ResultTest {
     void scoreTest() {
         Players players = new Players("pobi,jason", new Dealer());
         players.giveCards(new Deck());
-        GameResult.getPlayersCardsAndResult(players);
+        GameResultController.getPlayersCardsAndResult(players);
     }
 
     @Test
@@ -22,6 +22,6 @@ public class ResultTest {
         player.receiveCard(new Card(CardPattern.HEART, CardNumber.FIVE));
         dealer.receiveCard(new Card(CardPattern.SPADE, CardNumber.TEN));
         Players players = new Players(Collections.singletonList(player), dealer);
-        GameResult.getResult(players);
+        GameResultController.getResult(players);
     }
 }
